Fix soundness issues, move FFI closer to header, and expose as `mod sys`
The soundness is caused by the lack of an UnsafeCell around the
`os_unfair_lock` data, which *is* written to by `os_unfair_lock_lock`
and friends. This is the most important part.

I also brought the FFI bits in line with the actual header, and exposed
them in a `mod sys`.